refactor(ui): extract watcher hooks into app::watchers
