+++ src/ui/app/mod.rs
@@ -1,0 +1,6 @@
+mod watchers;
+
+pub use watchers::{
+    use_history_tracker, use_picker_reset_on_view, use_sync_watcher,
+    use_transcription_watcher,
+};
